little bugs fixed

+++ app/src/main/java/com/organiser/asyncTasks/TaskDeleter.java
@@ -33,6 +33,7 @@ public class TaskDeleter extends AsyncTask<Void,Void,Void> {
             }
         }
         try {
+            if(IDs.size()>0)
             taskService.deleteTask(IDs);
         } catch (Exception e) {
             e.printStackTrace();
